+++ docs/source/changelog.rst
@@ -12,6 +12,7 @@ Change Log
 - fixed unit tests to use unittest2 on Python 2.6
 - no longer supports jython 2.5. You'll have to upgrade to jython 2.7.
 - got rid of some byte/str handling cruft (because we no longer need to deal with jython 2.5)
+- implemented autoproxy support for serpent and json serializers. It is not possible to do this for marshal as well.
 
 
 **Pyro 4.20**